`MuchResult.for_boolean` -> `MuchResult.for`
This reworks the `.for_boolean` factory method to be a bit more
generic and just take any value and convert it to a boolean that
means success/failure. This expands the cases where this conversion
method can be used.

This also adds handling so that if `.for` is given a MuchResult,
it just updates its values and returns it.