add MuchResult#capture, MuchResult#capture!, and MuchResult.tap
The `#capture*` method allow for capturing child MuchResults and
composing them as part of a parent MuchResult. Said another way:
they allow you to aggregate MuchResults.

The `.tap` method is a convenience shortcut yielding a parent
result to a block so child results can be captured. It is a
macro for `MuchResult.new(...).tap { ... }`.