add MuchResult.transaction and MuchResult::Transaction handling
This API allows you to perform transactions on a transaction
receiver (e.g. `ActiveRecord::Base`) with MuchResult handling and
interactions.

MuchResult::Transaction is designed to be interacted with as
if it was a MuchResult. It delegates all MuchResult API calls to
its MuchResult. This allows you to capture sub-results in a
transaction and rollback if an exception occurs and/or a sub-result
failure.

The `MuchResult.transaction` method commits the transaction and
returns the transaction MuchResult if no exceptions occur.