remove MuchResult::Item
In building out the MuchResult API, it became apparent that having
an extra "item" object layer was pure overhead. An Item was
designed to be API-compatible with a MuchResult. MuchResult was
designed to contain either its Item or other MuchResults. However,
by moving the result SUCCESS/FAILURE directly on the MuchResult,
the need for tracking Item's goes away.

This also includes some README polish I noticed as I was reviewing
the API changes in the README.

This is just an internal implementation polish. There are no
overall behavior changes other than the `item` -> `result` renames.