rework capture! exception handling; prefer and `@data.exception`
This renames `result_exception` to `capture_exception` to use
more specific naming and to help avoid confusion.

This also switches to preferring any `@data.exception` that might
be set by the user when choosing the exception to raise on
`capture!`. This allows the user to get their own exceptions out
if they set them while falling back to some default if no
exception given.