add MuchResult#to_much_result; use in MuchResult.for
This switches `MuchResult.for` to check for a `to_much_result`
conversion method on the `value` and prefer using it if so. This
allows individual object types to each define how they are
converted to MuchResults and allows expanding better to new types.

This also adds a `#to_much_result` conversion method for
MuchResults themselves.