rework MuchResult#capture methods
This switches to the following #capture methods:
* `capture_for(...)` captures a given value converted to a result
* `capture_for!(...)` captures a given value converted to a result,
  raising an exception if the result is a failure
* `capture { ... }` captures a result returned by a block call
* `capture! { ... }` captures a result returned by a block call,
  raising an exception if the result is a failure

In addition, this adds `*_all` capture method. These methods, like
their `capture`/`capture!` equivalents, capture MuchResults.
However, these variants expect their results to be in an Array and
capture each result in that Array.

All of this is to provide a rich API for capturing sub-results
that can be used in a variety of different circumstances and
situations to ease using MuchResult.

Note: I also reworked the tests to group them logically so that
adding the new tests had some structure.